requirements.txt
fastapi
uvicorn
gradio
requests==2.31.0
urllib3==1.26.18
numpy
sentence-transformers
python-dotenv